everything_location v0.0.1 EverythingLocation

Everything Location is as SAAS that can verify the correctness of and address, update or correct values, and return a certainty of those values. This will also return a formatted address according to the rules of the country of that address. For more information see : https://www.everythinglocation.com

Example

iex> EverythingLocation.verify("Bastion 16, 3823 BP, Amersfoort, Netherlands")
%EverythingLocation.Result{
  administrative_area: "Utrecht",
  certainty_percentage: 100,
  city: "Amersfoort",
  country: "Netherlands",
  country_alpha2: "NL",
  country_alpha3: "NLD",
  formatted_address: "Bastion 16\n3823 BP  Amersfoort\nNetherlands",
  postal_code: "3823 BP",
  street: "Bastion 16",
  verified: true
]

Functions

verify(string)

Specs

verify(string) ::
  {:ok, %EverythingLocation.Result{__meta__: term, administrative_area: term, certainty_percentage: term, city: term, country: term, country_alpha2: term, country_alpha3: term, formatted_address: term, id: term, postal_code: term, street: term, verified: term}} |
  {:error, string}

verify(map) ::
  {:ok, %EverythingLocation.Result{__meta__: term, administrative_area: term, certainty_percentage: term, city: term, country: term, country_alpha2: term, country_alpha3: term, formatted_address: term, id: term, postal_code: term, street: term, verified: term}} |
  {:error, string}

Pass a Map of options to update/complete and verify its correctness. For accepted options, please refer to %EverythingLocation.Options{}. You must pass the options as a map, for example %{api_key: "my_api_key", address1: "my address"}