everything_location v0.0.1 EverythingLocation.Options

A struct for validating the parameters before passing them to the api

Functions

changeset(model, params)

Specs

changeset(%EverythingLocation.Options{__meta__: term, address1: term, address2: term, address3: term, address4: term, address5: term, address6: term, address7: term, address8: term, administrative_area: term, api_key: term, certify: term, country: term, enhance: term, geocode: term, id: term, locality: term, postal_code: term, suggest: term}, Map) :: %Ecto.Changeset{action: term, changes: term, constraints: term, errors: term, filters: term, model: term, optional: term, opts: term, params: term, prepare: term, repo: term, required: term, types: term, valid?: term, validations: term}

You can pass a Map with key values which will be cast and validated for use with the EverythingLocation API.

create(error)

Specs

create(%EverythingLocation.Options{__meta__: term, address1: term, address2: term, address3: term, address4: term, address5: term, address6: term, address7: term, address8: term, administrative_area: term, api_key: term, certify: term, country: term, enhance: term, geocode: term, id: term, locality: term, postal_code: term, suggest: term}) ::
  %{} |
  {:error, string}

Takes a EverythingLocation.Options struct and converts it to a map of key values formatted for the EverythingLocation API.